Eliminate base_command.py's import dependency on PipSession.
