+++ src/pip/_internal/commands/list.py
@@ -7,8 +7,8 @@ from pip._vendor import six
 from pip._vendor.six.moves import zip_longest
 
 from pip._internal.cli import cmdoptions
-from pip._internal.cli.base_command import Command
 from pip._internal.cli.cmdoptions import make_search_scope
+from pip._internal.cli.req_command import IndexGroupCommand
 from pip._internal.exceptions import CommandError
 from pip._internal.index import PackageFinder
 from pip._internal.models.selection_prefs import SelectionPreferences
@@ -21,7 +21,7 @@ from pip._internal.utils.packaging import get_installer
 logger = logging.getLogger(__name__)
 
 
-class ListCommand(Command):
+class ListCommand(IndexGroupCommand):
     """
     List installed packages, including editables.
